-> Added standalone runner for simulation -> Added multi-cluster geometry example

+++ simulator/runner/simulation_runner.py
@@ -10,27 +10,64 @@ from config import *
 
 class SimulationRunner:
 
-    def __init__(self, base_naming, geometry_infos, simulation_infos):
+    def __init__(self, base_naming, geometry_infos, simulation_infos=None):
         self._geometry_path = geometry_infos["file_path"]
         self._geometry_base_file = geometry_infos["base_file"]
         self._geometry_resolution = geometry_infos["resolution"]
         self._geometry_spacing = geometry_infos["spacing"]
-        self._number_of_maps = len(simulation_infos["compartment_ids"])
         self._base_naming = base_naming
-        self._simulation_path = simulation_infos["file_path"]
-        self._simulation_parameters = simulation_infos["param_file"]
-        self._compartment_ids = simulation_infos["compartment_ids"]
+        if simulation_infos:
+            self._number_of_maps = len(simulation_infos["compartment_ids"])
+            self._simulation_path = simulation_infos["file_path"]
+            self._simulation_parameters = simulation_infos["param_file"]
+            self._compartment_ids = simulation_infos["compartment_ids"]
+
+        self._run_simulation = True if simulation_infos else False
         self._event_loop = new_event_loop()
+
+    def run_simulation_standalone(self, output_folder, simulation_infos, test_mode=False):
+        simulation_output_folder = path.join(output_folder, "simulation_outputs")
+        geometry_output_folder = path.join(output_folder, "geometry_outputs")
+        singularity = path.join(singularity_path, singularity_name)
+
+        if not path.exists(simulation_output_folder):
+            makedirs(simulation_output_folder)
+
+        simulation_command = "singularity run -B {} --app launch_mitk {} -p {} -i {} -o {} {}".format(
+            ",".join([simulation_infos["file_path"], simulation_output_folder]),
+            singularity,
+            path.join(simulation_output_folder, "{}_simulation.ffp".format(self._base_naming)),
+            path.join(geometry_output_folder, self._base_naming) + "_merged_bundles.fib",
+            path.join(simulation_output_folder, self._base_naming),
+            "-v" if test_mode else ""
+        )
+
+        copyfile(
+            path.join(simulation_infos["file_path"], simulation_infos["param_file"]),
+            path.join(simulation_output_folder, "{}_simulation.ffp".format(self._base_naming))
+        )
+
+        set_event_loop(self._event_loop)
+        async_loop = get_event_loop()
+
+        with open(path.join(output_folder, "{}.log".format(self._base_naming)), "w+") as log_file:
+            print("Generating simulation geometry")
+            self._rename_and_copy_compartments(geometry_output_folder, simulation_output_folder)
+            print("Simulating DWI signal")
+            async_loop.run_until_complete(self._launch_command(simulation_command, log_file, "[RUNNING FIBERFOX]"))
+            async_loop.close()
 
     def run(self, output_folder, test_mode=False, relative_fiber_compartment=True):
         geometry_output_folder = path.join(output_folder, "geometry_outputs")
-        simulation_output_folder = path.join(output_folder, "simulation_outputs")
 
         if not path.exists(geometry_output_folder):
             makedirs(geometry_output_folder)
 
-        if not path.exists(simulation_output_folder):
-            makedirs(simulation_output_folder)
+        if self._run_simulation:
+            simulation_output_folder = path.join(output_folder, "simulation_outputs")
+
+            if not path.exists(simulation_output_folder):
+                makedirs(simulation_output_folder)
 
         singularity = path.join(singularity_path, singularity_name)
         geometry_command = "singularity run -B {} --app launch_voxsim {} -f {} -r {} -s {} -o {} --comp-map {} {} --quiet".format(
@@ -43,19 +80,21 @@ class SimulationRunner:
             "rel" if relative_fiber_compartment else "abs",
             "--no-process" if test_mode else "--quiet"
         )
-        simulation_command = "singularity run -B {} --app launch_mitk {} -p {} -i {} -o {} {}".format(
-            ",".join([self._simulation_path, simulation_output_folder]),
-            singularity,
-            path.join(simulation_output_folder, "{}_simulation.ffp".format(self._base_naming)),
-            path.join(geometry_output_folder, self._base_naming) + "_merged_bundles.fib",
-            path.join(simulation_output_folder, self._base_naming),
-            "-v" if test_mode else ""
-        )
 
-        copyfile(
-            path.join(self._simulation_path, self._simulation_parameters),
-            path.join(simulation_output_folder, "{}_simulation.ffp".format(self._base_naming))
-        )
+        if self._run_simulation:
+            simulation_command = "singularity run -B {} --app launch_mitk {} -p {} -i {} -o {} {}".format(
+                ",".join([self._simulation_path, simulation_output_folder]),
+                singularity,
+                path.join(simulation_output_folder, "{}_simulation.ffp".format(self._base_naming)),
+                path.join(geometry_output_folder, self._base_naming) + "_merged_bundles.fib",
+                path.join(simulation_output_folder, self._base_naming),
+                "-v" if test_mode else ""
+            )
+
+            copyfile(
+                path.join(self._simulation_path, self._simulation_parameters),
+                path.join(simulation_output_folder, "{}_simulation.ffp".format(self._base_naming))
+            )
 
         set_event_loop(self._event_loop)
         async_loop = get_event_loop()
@@ -63,10 +102,32 @@ class SimulationRunner:
         with open(path.join(output_folder, "{}.log".format(self._base_naming)), "w+") as log_file:
             print("Generating simulation geometry")
             async_loop.run_until_complete(self._launch_command(geometry_command, log_file, "[RUNNING VOXSIM]"))
-            self._rename_and_copy_compartments(geometry_output_folder, simulation_output_folder)
-            print("Simulating DWI signal")
-            async_loop.run_until_complete(self._launch_command(simulation_command, log_file, "[RUNNING FIBERFOX]"))
+            if self._run_simulation:
+                self._rename_and_copy_compartments(geometry_output_folder, simulation_output_folder)
+                print("Simulating DWI signal")
+                async_loop.run_until_complete(self._launch_command(simulation_command, log_file, "[RUNNING FIBERFOX]"))
             async_loop.close()
+
+    def _rename_and_copy_compartments_standalone(self, simulation_infos, geometry_output_folder, simulation_output_folder):
+        copyfile(
+            path.join(geometry_output_folder, self._base_naming + "0.nrrd"),
+            path.join(
+                simulation_output_folder,
+                "{}_simulation.ffp_VOLUME{}.nrrd".format(self._base_naming, simulation_infos["compartment_ids"][0])
+            )
+        )
+
+        if len(simulation_infos["compartment_ids"]) > 1:
+            copyfile(
+                path.join(geometry_output_folder, self._base_naming + "_mergedMaps.nrrd"),
+                path.join(
+                    simulation_output_folder,
+                    "{}_simulation.ffp_VOLUME{}.nrrd".format(self._base_naming, simulation_infos["compartment_ids"][1])
+                )
+            )
+
+        if len(simulation_infos["compartment_ids"]) > 2:
+            self._generate_background_map(geometry_output_folder, simulation_output_folder, simulation_infos["compartment_ids"])
 
     def _rename_and_copy_compartments(self, geometry_output_folder, simulation_output_folder):
         copyfile(
@@ -87,9 +148,9 @@ class SimulationRunner:
             )
 
         if len(self._compartment_ids) > 2:
-            self._generate_background_map(geometry_output_folder, simulation_output_folder)
+            self._generate_background_map(geometry_output_folder, simulation_output_folder, self._compartment_ids)
 
-    def _generate_background_map(self, geometry_output_folder, simulation_output_folder):
+    def _generate_background_map(self, geometry_output_folder, simulation_output_folder, compartment_ids):
         maps = [
             nrrd.read(
                 path.join(geometry_output_folder, "{}{}.nrrd".format(self._base_naming, 0))
@@ -106,7 +167,7 @@ class SimulationRunner:
         nrrd.write(
             path.join(
                 simulation_output_folder,
-                "{}_simulation.ffp_VOLUME{}.nrrd".format(self._base_naming, self._compartment_ids[-1])
+                "{}_simulation.ffp_VOLUME{}.nrrd".format(self._base_naming, compartment_ids[-1])
             ),
             extra_map,
             header
